backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
pydantic-settings==2.7.1
python-multipart==0.0.18
